Start `raspi-ws2811` driver
This driver is currently very simple and untested, since I forgot to get
a power supply for my LEDs, and I'm going on holiday for a week, so I
won't be able to test this driver properly for quite a while.

I want to allow customisation in the `build.rs` for this driver crate -
perhaps by some environment variables, or perhaps by a config file that
we read at compile-time - but I want a way for the user to choose high
or low speed mode for the lights, as well as which GPIO pin to use.
These features will come later.