Fix benchmarks
I'm not really sure if the benchmarks are actually fixed, because
they're now orders of magnitude faster than they were before. I don't
know if that's just because of the new effect architecture, or because
the benchmarks no longer have the overhead of an async runtime, or just
because I rewrote something and got it horribly wrong. But as far as I
can tell, the benchmarks are running just fine.